requirements.txt
streamlit
GitPython